requirements.txt
# Requirements for SelectAI project
--extra-index-url https://download.pytorch.org/whl/cu126
torch
teradataml
python-dotenv  
transformers
huggingface_hub           
accelerate
bitsandbytes
openpyxl
flask
pandas
